A2/Part B/microservice3/requirements.txt
Flask==2.0.2
requests==2.26.0
google-cloud-firestore
Flask-Cors==3.0.10
pytz